pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "ragconnect"
version = "0.1.0"
description = "LightRAG-based distributed memory system for AI team collaboration"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.27.0",
    "pydantic>=2.6.0",
    "pyyaml>=6.0.1",
    "mcp>=1.0.0",
    "click>=8.1.7",
    "anyio>=4.0.0",
    "neo4j>=5.28.0",
]

[project.scripts]
ragconnect-client = "client_gateway.mcp_server:main_sync"
ragconnect-server = "server_gateway.cli:cli"
ragconnect-web = "client_gateway.web_server:main_sync"
ragconnect-local-service = "client_gateway.local_service:cli"
ragconnect-host-helper = "server_gateway.host_helper:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["shared*", "client_gateway*", "server_gateway*", "local_embeddings*"]
